fix: wrap operations in transactions
To get a cursor to operate over a set of keys from an idb store, you
have to start a transaction. That transaction will remain open as long
as there are tasks in the microtask queue - when it empties the
transaction is automatically closed.

Transactions operate on an ObjectStore, puts and gets not to the object
store also seem to close the transaction.

This change adds a `_getObjectStore` method to the datastore which
creates a new transaction if there was no previous transation, or if
the previous transaction was closed.  All operations then take place as
part of this transaction.