feat!: convert to typescript
- Converts module to typescript
- Works around #6 by loading the query results in memory before yielding results
- Publishes as ESM-only
- Updates all deps to latest versions

The query performance will be bad, but at least it will be consistent.

Queries are only run by ipfs/helia during garbage collection which you don't normally do in the browser so it may not be as
bad as it seems.

BREAKING CHANGE: this module has been converted to typescript, updated to the latest interface-datastore version and is
ESM-only